Arquivo SwiftUI ScrollMovies adicionados em Components
